# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pet Shop",
        "description": "Pet shop API documentation",
        "contact": {
            "name": "Rabiul Islam",
            "email": "rir.cse.71@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/admin/create": {
            "post": {
                "description": "Create a new user",
                "operationId": "db08d61611cf4693bab87d7b6a3bceb6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "phone_number",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Mr."
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Admin"
                                    },
                                    "email": {
                                        "type": "email",
                                        "format": "string",
                                        "example": "user@email.com"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "+880175...."
                                    },
                                    "password": {
                                        "type": "password",
                                        "format": "string",
                                        "example": "*****"
                                    },
                                    "password_confirmation": {
                                        "type": "password",
                                        "format": "string",
                                        "example": "*****"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User created successfully"
                    },
                    "400": {
                        "description": "Something went wrong"
                    }
                }
            }
        },
        "/admin/user-listing": {
            "get": {
                "description": "User list",
                "operationId": "a26787adb1723341cb908a34cf6bfd4b",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "No. of page",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "How many record you want to get",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "first_name",
                        "in": "query",
                        "description": "First name of user",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User email",
                        "required": false,
                        "schema": {
                            "type": "email"
                        }
                    },
                    {
                        "name": "phone_number",
                        "in": "query",
                        "description": "User phone",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "address",
                        "in": "query",
                        "description": "User address",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "created_at",
                        "in": "query",
                        "description": "Create time",
                        "required": false,
                        "schema": {
                            "type": "datetime"
                        }
                    },
                    {
                        "name": "marketing",
                        "in": "query",
                        "description": "Is marketing or not",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User list"
                    },
                    "400": {
                        "description": "Something went wrong"
                    }
                }
            }
        },
        "/admin/user-edit/{uuid}": {
            "put": {
                "description": "Update an user",
                "operationId": "25d9e5ef7193a12fb97aaebff12a4792",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of User",
                        "required": true,
                        "schema": {
                            "type": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "phone_number",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Mr."
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Admin"
                                    },
                                    "email": {
                                        "type": "email",
                                        "format": "string",
                                        "example": "user@email.com"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "+880175...."
                                    },
                                    "password": {
                                        "type": "password",
                                        "format": "string",
                                        "example": "*****"
                                    },
                                    "password_confirmation": {
                                        "type": "password",
                                        "format": "string",
                                        "example": "*****"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully"
                    },
                    "400": {
                        "description": "Something went wrong"
                    }
                }
            }
        },
        "/admin/user-delete/{uuid}": {
            "delete": {
                "description": "Delete an user",
                "operationId": "b946acec5a3aab4f22a7d72aff5ef6b5",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of User",
                        "required": true,
                        "schema": {
                            "type": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully"
                    },
                    "400": {
                        "description": "Something went wrong"
                    }
                }
            }
        },
        "/admin/login": {
            "post": {
                "description": "Admin login",
                "operationId": "7a11fc783cfc18f5be41f3539a1e8121",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "format": "string",
                                        "example": "admin@email.com"
                                    },
                                    "password": {
                                        "type": "password",
                                        "format": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Logged in successfully"
                    },
                    "400": {
                        "description": "Something went wrong"
                    }
                }
            }
        },
        "/admin/logout": {
            "get": {
                "description": "Logout as an admin",
                "operationId": "54f1ea256aad39d9598e37b04423dab6",
                "responses": {
                    "200": {
                        "description": "Successfully logged out"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "User login",
                "operationId": "7be4b59b3e7b1ecc4e045533d8ee19bb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "format": "string",
                                        "example": "user@email.com"
                                    },
                                    "password": {
                                        "type": "password",
                                        "format": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Logged in successfully"
                    },
                    "400": {
                        "description": "Something went wrong"
                    }
                }
            }
        },
        "/user/logout": {
            "get": {
                "description": "Logout as an user",
                "operationId": "e36b977d1aa3dc11d6f28b2d24f5f469",
                "responses": {
                    "200": {
                        "description": "Successfully logged out"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
...
    }
}